Store prices as BigDecimals

+++ src/main/java/fi/xamk/tilavaraus/DemoApplication.java
@@ -7,6 +7,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 
 @SpringBootApplication
@@ -22,7 +23,7 @@ public class DemoApplication {
 		r.setId(id);
 		r.setName(name);
 		r.setCapacity(capacity);
-		r.setHourlyPrice(price);
+		r.setHourlyPrice(BigDecimal.valueOf(price));
 		return r;
 	}
